fix: autoscaling groups to NLBs attachment
The `target_group_arns` argument doesn't exist on eks managed node group, so we
need to handle the autoscaling groups attachment to NLBs in our module.

This commit uses a `nlbs_attachment` boolean argument of the node_groups to
attach the related autoscaling group to the load balancers. This works for both
types of node groups.